Break tests into their own file
